Unbreak make dest for FBSD
It turns out that we do need an afs.rc.fbsd that is set up for
transarc paths in this directory.  To get it to work properly
will require the user to symlink to it from a dir that gets
checked by rcorder, but them's the breaks.

Reviewed-on: http://gerrit.openafs.org/4378
Reviewed-by: Derrick Brashear <shadow@dementia.org>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
(cherry picked from commit 0b69fd6d5631482a786c2223a7dec0b2c9a07f92)

Change-Id: Ifd921019898954414a6d107702786c546d661bd4
Reviewed-on: http://gerrit.openafs.org/4855
Reviewed-by: Derrick Brashear <shadow@dementia.org>
Tested-by: BuildBot <buildbot@rampaginggeek.com>