viced: Remove old /vice/file/parms config file
This commit removes support for overriding command line options with
the contents of /vice/file/parms. This option has never been documented,
and only supports setting at most 15 command options. Replace the old
function with one which checks for the existence of this file, and
outputs a warning if it is found.

Change-Id: I933475c0eb31e78dbc8a9d31b2486aa14f57bfb0
Reviewed-on: http://gerrit.openafs.org/4716
Reviewed-by: Derrick Brashear <shadow@dementia.org>
Tested-by: Derrick Brashear <shadow@dementia.org>