libafs: Avoid using changing unixuser ticket data
PSetTokens was afs_osi_Alloc'ing after afs_osi_Free'ing the previous
token data. This can sleep, causing tu->stp to be pointing to garbage
while we wait to alloc. Additionally, rxkad_NewClientSecurityObject
can sleep while waiting to alloc memory, and so the given tu->stp
pointer given to it by afs_ConnBySA may be invalid by the time it
actually uses the data.

To fix this, we could implement unixuser locking to ensure mutual
exclusion of these events. However, this implements a more
conservative change for the 1.4 branch. In PSetTokens we alloc the new
memory before we change anything, and in afs_ConnBySA we make copies
of the ticket data before giving it to rxkad. With these changes, the
glock gives us enough serialization to avoid issues with tu->stp
changing underneath us.

This change is 1.4-specific. On the master branch, this issue is fixed
by implementing unixuser locks in change
Idd66d72f716b7e7dc08faa31ae43e9a23639bae3.

Change-Id: I3eb59e611531eca8105cf4d4f67eb1bb1196fd9c
Reviewed-on: http://gerrit.openafs.org/4649
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Derrick Brashear <shadow@dementia.org>